requirements.txt
pathlib
xxhash
ubelt
PyYaml